# docs/jsons/API_Documentation_1.0.0.swagger.json
{
  "openapi": "3.0.0",
  "tags": [],
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "contact": {},
    "description": ""
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/accounts": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string",
                  "required": true
                }
              }
            },
            "required": true
          }
        ],
        "tags": [
          "Accounts"
        ],
        "description": "create a user bank account on database.",
        "responses": {
          "201": {
            "description": "Returns created."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "409": {
            "description": "Returns an error for: user bank account already exists."
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "Accounts"
        ],
        "description": "searches for a user bank account on database.",
        "responses": {
          "200": {
            "description": "Returns a user bank account."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "404": {
            "description": "Returns an error for: user bank account not found."
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "Accounts"
        ],
        "description": "deletes a user bank account on database.",
        "responses": {
          "200": {
            "description": "Returns success for deleting user bank account."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "404": {
            "description": "Returns an error for: user bank account not found."
          }
        }
      }
    },
    "/accounts/{id}/increment": {
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "Accounts"
        ],
        "description": "increments user bank account balance on database.",
        "responses": {
          "200": {
            "description": "Returns success for user bank account balance increment."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "404": {
            "description": "Returns an error for: user bank account not found."
          }
        }
      }
    },
    "/authenticate": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "required": true
                },
                "password": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "tags": [
          "Authentication"
        ],
        "description": "Authenticate user by generating a JWT token.",
        "responses": {
          "200": {
            "description": "Returns created."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials."
          }
        }
      }
    },
    "/transfers": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senderAccountId": {
                  "type": "string",
                  "required": true
                },
                "receiverAccountId": {
                  "type": "string",
                  "required": true
                },
                "amount": {
                  "type": "number",
                  "required": true
                }
              }
            }
          }
        ],
        "tags": [
          "transfers"
        ],
        "description": "create a transfer on database.",
        "responses": {
          "201": {
            "description": "Returns created."
          },
          "400": {
            "description": "Returns an error for: user dont have founds."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "404": {
            "description": "Returns an error for: user bank account not found."
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transfers"
        ],
        "description": "list all transfers on database.",
        "responses": {
          "200": {
            "description": "Returns a list of transfers or an empty array."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          }
        }
      }
    },
    "/transfers/{accountId}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "transfers"
        ],
        "description": "list all transfers from a specific account.",
        "responses": {
          "200": {
            "description": "Returns a list of transfers or an empty array."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          }
        }
      }
    },
    "/users": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "tags": [
          "Users"
        ],
        "description": "list all users on database.",
        "responses": {
          "200": {
            "description": "Returns a list of users or an empty array."
          },
          "401": {
            "description": "Returns an error for: token not provided or invalid."
          }
        }
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "required": true
                },
                "email": {
                  "type": "string",
                  "required": true
                },
                "password": {
                  "type": "string",
                  "required": true
                },
                "birthdate": {
                  "type": "string",
                  "format": "date",
                  "required": true
                }
              }
            },
            "required": true
          }
        ],
        "tags": [
          "Users"
        ],
        "description": "create a user on database.",
        "responses": {
          "201": {
            "description": "Returns created."
          },
          "409": {
            "description": "Returns an error for: user already exists with same email."
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "Users"
        ],
        "description": "returns a unique user by id.",
        "responses": {
          "200": {
            "description": "Returns a user."
          },
          "401": {
            "description": "Returns an error for: token not provided or invalid."
          },
          "404": {
            "description": "Returns an error for: user not found."
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "required": true
                },
                "email": {
                  "type": "string",
                  "required": true
                },
                "password": {
                  "type": "string",
                  "required": true
                },
                "newPassword": {
                  "type": "string",
                  "required": false
                },
                "birthdate": {
                  "type": "string",
                  "format": "date",
                  "required": true
                }
              }
            },
            "required": true
          }
        ],
        "tags": [
          "Users"
        ],
        "description": "updates a user by id.",
        "responses": {
          "200": {
            "description": "Returns a user."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "404": {
            "description": "Returns an error for: user not found."
          },
          "409": {
            "description": "Returns an error for: user with sent email already registered."
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "Users"
        ],
        "description": "deletes a user by id.",
        "responses": {
          "200": {
            "description": "Returns deleted."
          },
          "401": {
            "description": "Returns an error for: user invalid credentials or token not provided/invalid."
          },
          "404": {
            "description": "Returns an error for: user not found."
          }
        }
      }
    }
  }
}